Minor refactoring
**Description:**
* Adjust the max idle connections in the connection pool of the
RestConnection object
* Remove the creation of multiple HubServices objects in the Initializer
constructor and the use of a random number to select one to use in the
thread
* Remove the multiple TaskExecutor bean definitions and place the bean
definition in a common location
* Use the start time and end times provided by the Execution objects in
the listener code of Phone Home
* Modified the return of _afterStep_ to _null_ so the step's exit status
is unchanged
* Removed multiple references to _EnableBatchProcessing_ annotation and
place a single reference in a common location
