+++ src/main/java/com/blackducksoftware/integration/fortify/batch/BatchSchedulerConfig.java
@@ -22,6 +22,7 @@
  */
 package com.blackducksoftware.integration.fortify.batch;
 
+import org.springframework.batch.core.configuration.annotation.EnableBatchProcessing;
 import org.springframework.batch.core.launch.JobLauncher;
 import org.springframework.batch.core.launch.support.SimpleJobLauncher;
 import org.springframework.batch.core.repository.JobRepository;
@@ -39,6 +40,7 @@ import org.springframework.transaction.PlatformTransactionManager;
  *
  */
 @Configuration
+@EnableBatchProcessing
 @EnableScheduling
 public class BatchSchedulerConfig {
